rename ImageCell to PhotoCell
