+++ src/main/resources/public/js/components/PhotoDisplay/PhotoCell.js
